+++ core/src/services/s3/backend.rs
@@ -1102,6 +1102,32 @@ impl Accessor for S3Backend {
             return Ok(RpStat::new(Metadata::new(EntryMode::DIR)));
         }
 
+        if path.ends_with('/') {
+            let resp = self
+                .core
+                .s3_list_objects(path, "", "", Some(1), None)
+                .await?;
+
+            let status = resp.status();
+
+            return match status {
+                StatusCode::OK => {
+                    let bs = resp.into_body().bytes().await?;
+                    let output: Output = quick_xml::de::from_reader(bs.reader())
+                        .map_err(new_xml_deserialize_error)?;
+                    if !output.contents.is_empty() {
+                        Ok(RpStat::new(Metadata::new(EntryMode::DIR)))
+                    } else {
+                        Err(
+                            Error::new(ErrorKind::NotFound, "The directory is not found")
+                                .with_context("path", path),
+                        )
+                    }
+                }
+                _ => Err(parse_error(resp).await?),
+            };
+        }
+
         let resp = self
             .core
             .s3_head_object(path, args.if_none_match(), args.if_match())
@@ -1111,9 +1137,6 @@ impl Accessor for S3Backend {
 
         match status {
             StatusCode::OK => parse_into_metadata(path, resp.headers()).map(RpStat::new),
-            StatusCode::NOT_FOUND if path.ends_with('/') => {
-                Ok(RpStat::new(Metadata::new(EntryMode::DIR)))
-            }
             _ => Err(parse_error(resp).await?),
         }
     }
